lifetimes: finished this section
I'm still not confident with lifetimes at all, and I'm not sure these
exercises are really representative of how they're used.

When calling a function that takes references, I don't see how adding
`<'a>` to the signature is any help to the compiler's borrow checker any
more than the fact the parameters were references to begin with. Doesn't
the borrow checker already enforce references to be valid? So when
calling a function with variables as arguments that themselves reference
some other data, wouldn't the borrow checker make sure those references
are valid up until then already?